Added: Parser
It took me more than 8 hours of work to get this far.
I still not sure if I should use the Crystal version of LIBSASS
and then use the DA_STYLE::Parser to sanitize values.
The point of this is to sanitize user input CSS to be executed
in the [domain].com instead of a member.domain.com sub-domain
like NeoCities or member.[domain].[different TLD] like Github .io
member pages.